Finished components abstract classes, implemented toString and equals method for this class. Also implemented the methods mentioned in CarFuelMonitorEnhanced: drive, refuel, isLowFuel.
